Move the LenChecker to its own file
